backend/pyproject.toml
[project]
name = "habit-tracker-backend"
version = "0.1.0"
description = "FastAPI backend for the Habit Tracker app"
requires-python = ">=3.9"

dependencies = [
    "fastapi",
    "pydantic",
    "sqlalchemy",
    "asyncpg",
    "python-jose",
    "passlib[argon2]>=1.7.4,<1.8",
    "argon2-cffi",
    "python-dotenv",
    "email-validator",
    "python-multipart",
    "uvicorn"
]

[project.optional-dependencies]
dev = [
    "pytest",
    "httpx",
    "pytest-asyncio",
    "aiosqlite",
    "greenlet",
]